Data.Algorithm.SatSolver.Lit

Type

data Lit a

Lit type

Constructors

Neg (Var a)

Pos (Var a)

Constructing

mkNeg :: Var a -> Lit a

mkNeg v makes a negative literal from propositionalvariable v.

>>> import qualified Data.Algoritm.SatSolver.Var as Var
>>> mkNeg (Var.mk "x1")
-"x1"

mkPos :: Var a -> Lit a

mkPos v makes a positive literal from propositional variable v.

>>> import qualified Data.Algoritm.SatSolver.Var as Var
>>> mkPos (Var.mk "x1")
+"x1"

mkNeg' :: a -> Lit a

mkNeg' n makes a propositional variable with name n and next makes a negative literal from this propositional variable.

>>> mkNeg' "x1"
-"x1"

mkPos' :: a -> Lit a

mkPos' n makes a propositional variable with name n and next makes a positive literal from this propositional variable.

>>> mkPos' "x1"
+"x1"

fromString :: String -> Lit String

fromString s makes a literal from a string. The first character must be either - or + (leading and trailing whitespaces are removed).

>>> fromString "-x1"
-"x1"
>>> :type fromString "-x1"
fromString "-x1" :: Lit String
>>> fromString "+x1"
+"x1"
>>> :type fromString "+x1"
fromString "+x1" :: Lit String

neg :: Lit a -> Lit a

neg l returns the opposite literal of literal l.

>>> neg $ mkNeg' "x1"
+"x1"
>>> neg $ mkNeg' "x1"
-"x1"

Transforming

getVar :: Lit a -> Var a

getVar l return the propositional variable literal l is defined on.

>>> getVar $ L.mkNeg' "x1"
"x1"
>>> getVar $ L.mkPos' "x1"
"x1"

toBool :: Lit a -> Bool

toBool l returns true if literal is positive and false otherwise.

>>> toBool  $ L.mkNeg' "x1"
False
>>> toBool  $ L.mkPos' "x1"
True